requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncmy==0.2.9
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic[email]==2.9.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
mercadopago==2.3.0
httpx==0.27.2
python-multipart==0.0.12
pytest==8.3.3
pytest-asyncio==0.24.0
